requirements/base.txt
psycopg2==2.8.5
Django==2.2.14
Pillow==7.1.2
django-crispy-forms==1.9.1
django-polymorphic==2.1.2
djangorestframework==3.11.0
django-sequences==2.5
django-braces==1.14.0
python-barcode==0.11.0
WeasyPrint==51
XlsxWriter==1.2.9
django-environ==0.4.5
django-countries==6.1.2
cie==0.208
django-waffle==1.0.0
django-import-export==2.3.0
django-reversion==3.0.7
